Uploading Python File
Should work. Bit worried it looks too similar to other people's examples, but I understand the process now. About the only flair I added on my own was adding "USD" to the Total Printups and all that. I just had fun assuming that the company might be a multi-national corporation or something.